requirements.txt
Flask~=2.3.2
python-dotenv~=1.0.0
pymongo[srv]~=4.4.0
Flask-PyMongo
Werkzeug~=2.3.6